pyproject.toml
[tool.poetry]
name = "voice-to-text-demo"
version = "0.1.0"
description = "Voice to text app built with Django"
authors = ["HeskethGD <graham.hesketh@opsmorph.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.3"
djangorestframework = "^3.15.2"
drf-spectacular = "^0.27.2"
openai = "^1.54.4"
python-dotenv = "^1.0.1"
streamlit = "^1.40.1"
requests = "^2.32.3"
audio-recorder-streamlit = "^0.0.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
